@@ -1,26 +1,26 @@
 import { Injectable } from '@angular/core';
 
-import {Http ,Response} from '@angular/http';
+// import {Http ,Response} from '@angular/http';
 import 'rxjs/add/operator/toPromise';
 
-import {Observable} from 'rxjs/Observable';
+// import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 
 @Injectable()
 export class MiHttpService {
   
-  constructor(public http:Http) { }
+  constructor(){}//public http:Http) { }
   
-  public httpGetPromise(url: string, objeto:any){
+  // public httpGetPromise(url: string, objeto:any){
 
 
-    return this.http
-    .get(url)
-    .toPromise()
-    .then(this.extraerDatos)
-    .catch(this.handleError);
-  }
+  //   return this.http
+  //   .get(url)
+  //   .toPromise()
+  //   .then(this.extraerDatos)
+  //   .catch(this.handleError);
+  // }
 
   private extraerDatos(resp:Response) {
 
