@@ -1,43 +1,43 @@
 import { log } from 'util';
 import { Injectable } from '@angular/core';
 
-import { Http, Response } from '@angular/http';
+// import { Http, Response } from '@angular/http';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/toPromise';
+// import { Observable } from 'rxjs/Observable';
+// import 'rxjs/add/operator/catch';
+// import 'rxjs/add/operator/map';
+// import 'rxjs/add/operator/toPromise';
 
 @Injectable()
 export class MiHttpService {
 
-  constructor( public http: Http ) { }
+  constructor( ){} //public http: Http ) { }
 
-  public httpGetP ( url: string)
-  {
-    return this.http
-    .get( url )
-    .toPromise()
-    .then( this.extractData )
-    .catch( this.handleError );
-  }
+  // public httpGetP ( url: string)
+  // {
+  //   return this.http
+  //   .get( url )
+  //   .toPromise()
+  //   .then( this.extractData )
+  //   .catch( this.handleError );
+  // }
 
-  public httpPostP( url: string, objeto: any )
-  {
-    return this.http
-    .get( url )
-    .subscribe( data => {
-      console.log( data );
-      return data;
-    });
-  }
+  // public httpPostP( url: string, objeto: any )
+  // {
+  //   return this.http
+  //   .get( url )
+  //   .subscribe( data => {
+  //     console.log( data );
+  //     return data;
+  //   });
+  // }
 
-  public httpGetO ( url: string): Observable<Response>
-  {
-    return this.http.get( url )
-      .map( ( res: Response ) => res.json())
-      .catch( ( err: any ) => Observable.throw(err.json().error || 'Server error'));
-  }
+  // public httpGetO ( url: string): Observable<Response>
+  // {
+  //   return this.http.get( url )
+  //     .map( ( res: Response ) => res.json())
+  //     .catch( ( err: any ) => Observable.throw(err.json().error || 'Server error'));
+  // }
 
 
   private extractData ( res: Response )
