@@ -21,27 +21,27 @@ export class JugadoresListadoComponent implements OnInit {
   }
 
 
-  TraerTodos(){
-    //alert("totos");
-    this.miJugadoresServicio.traertodos('jugadores/','todos').then(data=>{
-      //console.info("jugadores listado",(data));
-      this.listado= data;
+  // TraerTodos(){
+  //   //alert("totos");
+  //   this.miJugadoresServicio.traertodos('jugadores/','todos').then(data=>{
+  //     //console.info("jugadores listado",(data));
+  //     this.listado= data;
 
-    })
-  }
-  TraerGanadores(){
-    this.miJugadoresServicio.traertodos('jugadores/','ganadores').then(data=>{
-      //console.info("jugadores listado",(data));
-      this.listado= data;
+  //   })
+  // }
+  // TraerGanadores(){
+  //   this.miJugadoresServicio.traertodos('jugadores/','ganadores').then(data=>{
+  //     //console.info("jugadores listado",(data));
+  //     this.listado= data;
 
-    })
-  }
-  TraerPerdedores(){
-    this.miJugadoresServicio.traertodos('jugadores/','perdedores').then(data=>{
-      //console.info("jugadores listado",(data));
-      this.listado= data;
+  //   })
+  // }
+  // TraerPerdedores(){
+  //   this.miJugadoresServicio.traertodos('jugadores/','perdedores').then(data=>{
+  //     //console.info("jugadores listado",(data));
+  //     this.listado= data;
 
-    })
-  }
+  //   })
+  // }
 
 }
