@@ -14,11 +14,11 @@ export class ListadoDePaisesComponent implements OnInit {
    }
 
   ngOnInit() {
-    this.miServicioDePaises.listar()
-    .then(datos=>{
-      console.info("listado de paises",datos);
-      this.listadoDePaises=datos;
-    });
+    // this.miServicioDePaises.listar()
+    // .then(datos=>{
+    //   console.info("listado de paises",datos);
+    //   this.listadoDePaises=datos;
+    // });
   }
 
 }
