@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 
 import {Subscription} from "rxjs";
-import {TimerObservable} from "rxjs/observable/TimerObservable";
+// import {TimerObservable} from "rxjs/observable/TimerObservable";
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
@@ -36,46 +36,46 @@ export class LoginComponent implements OnInit {
       this.router.navigate(['/Principal']);
     }
   }
-  MoverBarraDeProgreso() {
+  // MoverBarraDeProgreso() {
     
-    this.logeando=false;
-    this.clase="progress-bar progress-bar-danger progress-bar-striped active";
-    this.progresoMensaje="NSA spy..."; 
-    let timer = TimerObservable.create(200, 50);
-    this.subscription = timer.subscribe(t => {
-      console.log("inicio");
-      this.progreso=this.progreso+1;
-      this.ProgresoDeAncho=this.progreso+20+"%";
-      switch (this.progreso) {
-        case 15:
-        this.clase="progress-bar progress-bar-warning progress-bar-striped active";
-        this.progresoMensaje="Verificando ADN..."; 
-          break;
-        case 30:
-          this.clase="progress-bar progress-bar-Info progress-bar-striped active";
-          this.progresoMensaje="Adjustando encriptación.."; 
-          break;
-          case 60:
-          this.clase="progress-bar progress-bar-success progress-bar-striped active";
-          this.progresoMensaje="Recompilando Info del dispositivo..";
-          break;
-          case 75:
-          this.clase="progress-bar progress-bar-success progress-bar-striped active";
-          this.progresoMensaje="Recompilando claves facebook, gmail, chats..";
-          break;
-          case 85:
-          this.clase="progress-bar progress-bar-success progress-bar-striped active";
-          this.progresoMensaje="Instalando KeyLogger..";
-          break;
+  //   this.logeando=false;
+  //   this.clase="progress-bar progress-bar-danger progress-bar-striped active";
+  //   this.progresoMensaje="NSA spy..."; 
+  //   let timer = TimerObservable.create(200, 50);
+  //   this.subscription = timer.subscribe(t => {
+  //     console.log("inicio");
+  //     this.progreso=this.progreso+1;
+  //     this.ProgresoDeAncho=this.progreso+20+"%";
+  //     switch (this.progreso) {
+  //       case 15:
+  //       this.clase="progress-bar progress-bar-warning progress-bar-striped active";
+  //       this.progresoMensaje="Verificando ADN..."; 
+  //         break;
+  //       case 30:
+  //         this.clase="progress-bar progress-bar-Info progress-bar-striped active";
+  //         this.progresoMensaje="Adjustando encriptación.."; 
+  //         break;
+  //         case 60:
+  //         this.clase="progress-bar progress-bar-success progress-bar-striped active";
+  //         this.progresoMensaje="Recompilando Info del dispositivo..";
+  //         break;
+  //         case 75:
+  //         this.clase="progress-bar progress-bar-success progress-bar-striped active";
+  //         this.progresoMensaje="Recompilando claves facebook, gmail, chats..";
+  //         break;
+  //         case 85:
+  //         this.clase="progress-bar progress-bar-success progress-bar-striped active";
+  //         this.progresoMensaje="Instalando KeyLogger..";
+  //         break;
           
-        case 100:
-          console.log("final");
-          this.subscription.unsubscribe();
-          this.Entrar();
-          break;
-      }     
-    });
+  //       case 100:
+  //         console.log("final");
+  //         this.subscription.unsubscribe();
+  //         this.Entrar();
+  //         break;
+  //     }     
+  //   });
     //this.logeando=true;
-  }
+  //}
 
 }
