@@ -8,18 +8,18 @@ export class JuegoServiceService {
 
   peticion:any;
   constructor( public miHttp: MiHttpService ) {
-    this.peticion = this.miHttp.httpGetO("http://localhost:3003");
+    // this.peticion = this.miHttp.httpGetO("http://localhost:3003");
 //    this.peticion = this.miHttp.httpGetO("https://restcountries.eu/rest/v2/all");
   }
 
   public listar(): Array<Juego> {
-   this.miHttp.httpGetP("https://restcountries.eu/rest/v2/all")
-    .then( data => {
-      console.log( data );
-    })
-    .catch( err => {
-      console.log( err );
-    });
+  //  this.miHttp.httpGetP("https://restcountries.eu/rest/v2/all")
+  //   .then( data => {
+  //     console.log( data );
+  //   })
+  //   .catch( err => {
+  //     console.log( err );
+  //   });
    
   
     this.peticion
