@@ -6,16 +6,16 @@ export class PaisesService {
   constructor(public miHttp: MiHttpService ) { }
 
 
-  public listar():Promise<Array<any>> {
-       return   this.miHttp.httpGetP("https://restcountries.eu/rest/v2/all")
-          .then( data => {
-            console.log( data );
-            return data;
-          })
-          .catch( err => {
-            console.log( err );
-            return null;
-          });
-          //return null;
-    }
+  // public listar():Promise<Array<any>> {
+  //      return   this.miHttp.httpGetP("https://restcountries.eu/rest/v2/all")
+  //         .then( data => {
+  //           console.log( data );
+  //           return data;
+  //         })
+  //         .catch( err => {
+  //           console.log( err );
+  //           return null;
+  //         });
+  //         //return null;
+  //   }
 }
